requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.4
sqlalchemy==2.0.25
streamlit==1.31.0
plotly==5.18.0
PyJWT==2.8.0
cryptography==41.0.7
pytest==7.4.4